move code that generate unit tests and docs to main so that extensions can easily extend it
